moved types into the right folder
