+++ core/terraform/platform.go
@@ -1,0 +1,4 @@
+package terraform
+
+// Platform that relates to this functionality
+const Platform = "terraform"
